requirements.txt
falkordb
langchain
langchain-openai
langgraph
python-dotenv
pymupdf
pydantic
rapidfuzz
unstructured
streamlit
pyyaml

